RavenDB-19358: PR comments
